+++ sdk/core/azure-core/inc/azure/core/internal/strings.hpp
@@ -38,10 +38,12 @@ namespace Azure { namespace Core { namespace _internal {
       return IsDigit(c) || (c >= 'A' && c <= 'F') || (c >= 'a' && c <= 'f');
     }
 
-    static constexpr bool IsAlphaNumeric(char c) noexcept
+    static constexpr bool IsAlpha(char c) noexcept
     {
-      return IsDigit(c) || (c >= 'A' && c <= 'Z') || (c >= 'a' && c <= 'z');
+      return (c >= 'A' && c <= 'Z') || (c >= 'a' && c <= 'z');
     }
+
+    static constexpr bool IsAlphaNumeric(char c) noexcept { return IsDigit(c) || IsAlpha(c); }
 
     static constexpr bool IsSpace(char c) noexcept { return c == ' ' || (c >= '\t' && c <= '\r'); }